Add LineExtra with start and end anchor chars
